I created the IRFormFieldCell class and added row selection behavior to the form table.
I created the IRFormFieldCell class to act as the base class for all
form field cell views. I modified the IRForm class to handle the event
when the user selects a row in the form. IRForm will now invoke the
activateCell method on the IRFormFieldCell object to do something. I
implemented the activateCell method on the IRTextFormField class to
make the text field the first responder when the cell is selected.